use context provider with contacts for the whole app
