+++ src/app/_components/footer/footer.tsx
@@ -1,10 +1,9 @@
 'use client';
+
 import { MapPinIcon, PhoneIcon } from '@heroicons/react/24/outline';
 import classes from './footer.module.scss';
 
-import { useContext } from 'react';
-import { useContacts } from '@/app/_lib/hooks';
-import AppContextProvider, { AppContext } from '@/app/_lib/context';
+import { useAppContext } from '../context/context';
 
 interface FooterParams {
   propClasses?: string;
@@ -12,22 +11,18 @@ interface FooterParams {
 
 export default function Footer(params: FooterParams) {
   const { propClasses = '' } = params;
-  const { contacts: contactsFromContext } = useContext(AppContext);
-  // const { contacts } = useContacts();
-  console.log('contactsFromContext', contactsFromContext);
+  const { contacts } = useAppContext();
 
   return (
-    <AppContextProvider>
-      <div className={`${classes.footer} ${propClasses}`}>
-        <p>
-          <MapPinIcon className="h-6 w-6" />
-          {contactsFromContext[0]?.address}
-        </p>
-        <p>
-          <PhoneIcon className="h-6 w-6" />
-          {contactsFromContext[0]?.phone}
-        </p>
-      </div>
-    </AppContextProvider>
+    <div className={`${classes.footer} ${propClasses}`}>
+      <p>
+        <MapPinIcon className="h-6 w-6" />
+        {contacts[0]?.address}
+      </p>
+      <p>
+        <PhoneIcon className="h-6 w-6" />
+        {contacts[0]?.phone}
+      </p>
+    </div>
   );
 }
