+++ src/app/contacts/page.tsx
@@ -1,3 +1,9 @@
+'use client';
+
+import { useAppContext } from '../_components/context/context';
+
 export default function Contacts() {
+  const { contacts } = useAppContext();
+
   return <main>Contacts</main>;
 }
